Update the site styling
I felt that the styling could be improved just that much more,
specifically the size of the fonts for the input fields. Technically,
not necessary, but I felt by making the fonts a bit bigger, they'd fit
the theme of the site better and be that much more interesting, fun, and
relaxed in nature.