Remove the Halloween images from the repository
The images are now stored in a DigitalOcean Space, partly as they need
to be publicly accessible so that Twilio can access them and send them
as part of the SMS, and partly because it's not efficient to store
images in a Git repository.

See https://www.perforce.com/blog/vcs/how-git-lfs-works for more
information on not storing images in Git.

Given both of these reasons, this change removes the images from the
repository.